removes NormalyzerDE and yarrr deps, fixes normalization bug
Fixes important bug in quantile normalization that cause log2Norm, rlrNorm, and giNorm to be incorrect.

Rewrote the giNorm and rlrNorm functions to no longer rely on the NormalyzerDE functions we were using to do these normalizations. For giNorm, new version is also much faster.

Removed yarrr, since we only used it for color palettes in what is, as far as I can tell, is a now orphan function.

Overall, went from 181 dependencies to 101.