adds initial limma csv table saving
Currently, has summary, big query, combined results, and per contrast results. Need to deal with spreadsheet.

Add in a new default file structure for this. Probably need to consult with others to determine a final structure for projects like this.

Passes the data through the fit_limma_model function, to make things a little easier for on the end user when writing logs.