fix a bug when user specified table_columns with bad JS colnames
A bug that would occur when users specified table columns (or a title column that gets added to the table) that contained names incompatible with JS. I had fixed this issue with PR 192, but didn't check the outputs closely enough: it only worked for the first report. Something was getting overwritten in the later reports for the loop., though print debugging indicated it was all fine. In any case, moving the colname correction outside the loop and differentiating between the original and fixed colnames did the trick.